+++ iced_exwlshell/src/lib.rs
@@ -15,12 +15,17 @@ pub mod reexport {
     pub use exwlshellev::NewInputPanelSettings;
     pub use exwlshellev::NewLayerShellSettings;
     pub use exwlshellev::OutputOption;
+    pub use exwlshellev::PopupPlacement;
     pub use exwlshellev::WithConnection;
     pub use exwlshellev::WlShellType;
     pub use exwlshellev::reexport::Anchor;
     pub use exwlshellev::reexport::KeyboardInteractivity;
     pub use exwlshellev::reexport::Layer;
     pub use exwlshellev::reexport::wayland_client::{WlRegion, wl_keyboard};
+    pub use exwlshellev::reexport::xdg_positioner::{
+        Anchor as PopupAnchor, ConstraintAdjustment as PopupConstraintAdjustment,
+        Gravity as PopupGravity,
+    };
     pub mod core {
         pub use iced_core::*;
     }