+++ iced_exwlshell_macros/src/lib.rs
@@ -90,19 +90,19 @@ pub fn to_exwlshell_message(
                     )
 
                 }
-                fn base_window_open(settings: iced_exwlshell::actions::IcedXdgWindowSettings) -> (iced_exwlshell::reexport::IcedId, iced_exwlshell::reexport::Task<Self>) {
-                    let id = iced_exwlshell::reexport::IcedId::unique();
-                    (
-                        id,
-                        iced_exwlshell::reexport::Task::done(Self::NewBaseWindow { settings, id })
-                    )
-
-                }
                 fn menu_open(settings: iced_exwlshell::actions::IcedNewMenuSettings) -> (iced_exwlshell::reexport::IcedId, iced_exwlshell::reexport::Task<Self>) {
                     let id = iced_exwlshell::reexport::IcedId::unique();
                     (
                         id,
                         iced_exwlshell::reexport::Task::done(Self::NewMenu { settings, id })
+                    )
+
+                }
+                fn base_window_open(settings: iced_exwlshell::actions::IcedXdgWindowSettings) -> (iced_exwlshell::reexport::IcedId, iced_exwlshell::reexport::Task<Self>) {
+                    let id = iced_exwlshell::reexport::IcedId::unique();
+                    (
+                        id,
+                        iced_exwlshell::reexport::Task::done(Self::NewBaseWindow { settings, id })
                     )
 
                 }
@@ -137,7 +137,7 @@ pub fn to_exwlshell_message(
                         Self::NewLayerShell {settings, id } => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewLayerShell { settings, id })),
                         Self::NewBaseWindow {settings, id } => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewBaseWindow { settings, id })),
                         Self::NewPopUp { settings, id } => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewPopUp { settings, id })),
-                        Self::NewMenu { settings, id } =>  Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewMenu {settings, id })),
+                        Self::NewMenu { settings, id } => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewMenu { settings, id })),
                         Self::NewInputPanel {settings, id } => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::NewInputPanel { settings, id })),
                         Self::RemoveWindow(id) => Ok(ExwlShellCustomActionWithId::new(Some(id), ExwlShellCustomAction::RemoveWindow)),
                         Self::ForgetLastOutput => Ok(ExwlShellCustomActionWithId::new(None, ExwlShellCustomAction::ForgetLastOutput)),