+++ iced_layershell_macros/src/lib.rs
@@ -76,19 +76,19 @@ pub fn to_layer_message(attr: TokenStream2, input: TokenStream2) -> manyhow::Res
                         )
 
                     }
-                    fn base_window_open(settings: iced_layershell::actions::IcedXdgWindowSettings) -> (iced_layershell::reexport::IcedId, iced_layershell::reexport::Task<Self>) {
-                        let id = iced_layershell::reexport::IcedId::unique();
-                        (
-                            id,
-                            iced_layershell::reexport::Task::done(Self::NewBaseWindow { settings, id })
-                        )
-
-                    }
                     fn menu_open(settings: iced_layershell::actions::IcedNewMenuSettings) -> (iced_layershell::reexport::IcedId, iced_layershell::reexport::Task<Self>) {
                         let id = iced_layershell::reexport::IcedId::unique();
                         (
                             id,
                             iced_layershell::reexport::Task::done(Self::NewMenu { settings, id })
+                        )
+
+                    }
+                    fn base_window_open(settings: iced_layershell::actions::IcedXdgWindowSettings) -> (iced_layershell::reexport::IcedId, iced_layershell::reexport::Task<Self>) {
+                        let id = iced_layershell::reexport::IcedId::unique();
+                        (
+                            id,
+                            iced_layershell::reexport::Task::done(Self::NewBaseWindow { settings, id })
                         )
 
                     }
@@ -116,7 +116,7 @@ pub fn to_layer_message(attr: TokenStream2, input: TokenStream2) -> manyhow::Res
                             Self::NewLayerShell {settings, id } => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewLayerShell { settings, id })),
                             Self::NewBaseWindow {settings, id } => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewBaseWindow { settings, id })),
                             Self::NewPopUp { settings, id } => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewPopUp { settings, id })),
-                            Self::NewMenu { settings, id } =>  Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewMenu {settings, id })),
+                            Self::NewMenu { settings, id } => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewMenu { settings, id })),
                             Self::NewInputPanel {settings, id } => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::NewInputPanel { settings, id })),
                             Self::RemoveWindow(id) => Ok(LayerShellCustomActionWithId::new(Some(id), LayerShellCustomAction::RemoveWindow)),
                             Self::ForgetLastOutput => Ok(LayerShellCustomActionWithId::new(None, LayerShellCustomAction::ForgetLastOutput)),